CrystallBall code...
Arduino is an open source project. Here the project was built with the Arduino programming language(based on Wiring) and the Arduino Software IDE based on Processing. The pinModes( ) are set and the LCD set to output Alphanumeric characters of our choice. A switch statement is used to execute different lines of code depending on a chosen number. This number is randomly generated through the random( ) function. Random will generate a random number between 0 and 1. Random(8) will extend that range to 0 to 7.